Move compress package
