+++ proxy/src/main/java/net/md_5/bungee/BungeeCord.java
@@ -71,7 +71,7 @@ import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Plugin;
 import net.md_5.bungee.api.plugin.PluginManager;
 import net.md_5.bungee.command.*;
-import net.md_5.bungee.compress.CompressFactory;
+import net.md_5.bungee.netty.compress.CompressFactory;
 import net.md_5.bungee.conf.YamlConfig;
 import net.md_5.bungee.forge.ForgeConstants;
 import net.md_5.bungee.log.LoggingOutputStream;
