+++ proxy/src/main/java/net/md_5/bungee/netty/ChannelWrapper.java
@@ -1,7 +1,7 @@
 package net.md_5.bungee.netty;
 
-import net.md_5.bungee.compress.PacketCompressor;
-import net.md_5.bungee.compress.PacketDecompressor;
+import net.md_5.bungee.netty.compress.PacketCompressor;
+import net.md_5.bungee.netty.compress.PacketDecompressor;
 import net.md_5.bungee.protocol.PacketWrapper;
 import com.google.common.base.Preconditions;
 import io.netty.channel.Channel;
